feat: update dependencies to polkadot v1 release

+++ crates/annuity/src/tests.rs
@@ -8,19 +8,31 @@ fn should_calculate_emission_rewards() {
     run_test(|| {
         <Balances as Currency<AccountId>>::make_free_balance_be(&Annuity::account_id(), YEAR_1_REWARDS);
         Annuity::update_reward_per_block();
-        assert_eq!(Annuity::reward_per_block(), YEAR_1_REWARDS / EmissionPeriod::get());
+        assert_eq!(
+            Annuity::reward_per_block(),
+            YEAR_1_REWARDS / EmissionPeriod::get() as u128
+        );
 
         <Balances as Currency<AccountId>>::make_free_balance_be(&Annuity::account_id(), YEAR_2_REWARDS);
         Annuity::update_reward_per_block();
-        assert_eq!(Annuity::reward_per_block(), YEAR_2_REWARDS / EmissionPeriod::get());
+        assert_eq!(
+            Annuity::reward_per_block(),
+            YEAR_2_REWARDS / EmissionPeriod::get() as u128
+        );
 
         <Balances as Currency<AccountId>>::make_free_balance_be(&Annuity::account_id(), YEAR_3_REWARDS);
         Annuity::update_reward_per_block();
-        assert_eq!(Annuity::reward_per_block(), YEAR_3_REWARDS / EmissionPeriod::get());
+        assert_eq!(
+            Annuity::reward_per_block(),
+            YEAR_3_REWARDS / EmissionPeriod::get() as u128
+        );
 
         <Balances as Currency<AccountId>>::make_free_balance_be(&Annuity::account_id(), YEAR_4_REWARDS);
         Annuity::update_reward_per_block();
-        assert_eq!(Annuity::reward_per_block(), YEAR_4_REWARDS / EmissionPeriod::get());
+        assert_eq!(
+            Annuity::reward_per_block(),
+            YEAR_4_REWARDS / EmissionPeriod::get() as u128
+        );
     })
 }
 
@@ -29,7 +41,10 @@ fn should_set_reward_per_wrapped() {
     run_test(|| {
         <Balances as Currency<AccountId>>::make_free_balance_be(&Annuity::account_id(), YEAR_1_REWARDS);
         Annuity::update_reward_per_block();
-        assert_eq!(Annuity::min_reward_per_block(), YEAR_1_REWARDS / EmissionPeriod::get());
+        assert_eq!(
+            Annuity::min_reward_per_block(),
+            YEAR_1_REWARDS / EmissionPeriod::get() as u128
+        );
         let reward_per_wrapped = 100;
         assert_ok!(Annuity::set_reward_per_wrapped(
             RuntimeOrigin::root(),
